requirements.txt
# Core
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Telegram
telethon==1.34.0

# AI
openai==1.12.0
pinecone-client==3.0.0

# Templates
jinja2==3.1.3
aiofiles==23.2.1

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0

# Scheduler
apscheduler==3.10.4

# Optional: Supabase (for media storage)
supabase==2.3.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.5
